src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/host-pseudo-class-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
Test whether :host matches a shadow host correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

:host out of shadow tree should not match any shadow hosts.
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
